Extract pure matchTeam and unit-test it
Split the team-hint matching (slug-or-name, case-insensitive, note on miss)
out of the network-bound resolveTeamSlug into a pure matchTeam in job-utils,
mirroring the functional-core/imperative-shell shape of the other helpers.
One no-mock unit test covers the branches that actually regress; the network
contract stays with the QA pass.
