+++ src/tools/design-import/job-utils.ts
@@ -40,6 +40,23 @@ export function slugify(title?: string): string | undefined {
   return slug || undefined;
 }
 
+export type TeamRef = { slug?: string; name?: string };
+
+// Matches a caller-supplied team hint against the user's teams, on slug or name
+// (the caller may pass either), case-insensitively. Returns the canonical slug
+// when found; otherwise an explanatory note so the deploy can fall back to the
+// default team and still tell the caller what happened, instead of failing.
+export function matchTeam(teams: TeamRef[], requested: string): { slug?: string; note?: string } {
+  const wanted = requested.toLowerCase();
+  const match = teams.find((team) => team.slug?.toLowerCase() === wanted || team.name?.toLowerCase() === wanted);
+  if (match?.slug) {
+    return { slug: match.slug };
+  }
+  return {
+    note: `Team "${requested}" was not found in your Netlify teams, so the design was deployed to your default team. Tell me the exact team name if you want it moved.`,
+  };
+}
+
 export function encodeJob(siteId: string, deployId: string): string {
   return Buffer.from(`${siteId}:${deployId}`).toString('base64url');
 }
